Refactoring
General code maintenance and cleanup
