+++ cypress/e2e/functionality/01-login.cy.ts
@@ -1,4 +1,4 @@
-import { goodPassword, goodUser } from "../../support/commands";
+import { goodPassword, goodUser } from "../../support/constants";
 
 describe('Login Tests', () => {
   const errorText = "The username or password was not recognised";
