+++ cypress/e2e/functionality/05-file-upload.cy.ts
@@ -1,4 +1,4 @@
-import { goodPassword, goodUser } from "../../support/commands";
+import { goodPassword, goodUser } from "../../support/constants";
 
 describe('Files upload', () => {
     beforeEach(() => {
