+++ cypress/e2e/functionality/07-layouts.cy.ts
@@ -1,6 +1,7 @@
 import { LUA } from "../../../src/frontend/js/lib/util/formatters/lua";
-import { goodPassword, goodUser } from "../../support/commands";
-import { ICodeLayoutBuilder, ICurvalLayoutBuilder, IDropdownLayoutBuilder, ILayoutBuilder, LayoutBuilder } from "../../support/LayoutBuilder";
+import { LayoutBuilder } from "../../support/builders/layout/LayoutBuilder";
+import { ICodeLayoutBuilder, ICurvalLayoutBuilder, IDropdownLayoutBuilder, ILayoutBuilder } from "../../support/builders/layout/interfaces";
+import { goodPassword, goodUser } from "../../support/constants";
 
 describe("Layout creation tests", () => {
     const refShortName = "tr";
@@ -12,7 +13,7 @@ describe("Layout creation tests", () => {
     context("Simple layout creation", () => {
         ["TEXT", "INTEGER", "DATE", "DATE-RANGE", "DOCUMENT", "PERSON", "TREE"].forEach((layoutType) => {
             it(`Should create ${layoutType} layout`, () => {
-                const builder = <ILayoutBuilder>LayoutBuilder.create(<any>layoutType);
+                const builder:ILayoutBuilder = LayoutBuilder.create(<any>layoutType) as ILayoutBuilder;
                 builder
                     .withName("test1")
                     .withShortName("t1");
@@ -30,7 +31,7 @@ describe("Layout creation tests", () => {
         it("Should error on creating a RAG layout with invalid code", () => {
             const refField = LayoutBuilder.create("TEXT");
             refField.withName("refField").withShortName(refShortName);
-            const builder = <ICodeLayoutBuilder>LayoutBuilder.create("RAG");
+            const builder:ICodeLayoutBuilder = LayoutBuilder.create("RAG") as ICodeLayoutBuilder;
             builder
                 .withName("test1")
                 .withShortName("t1")
@@ -58,7 +59,7 @@ end
         it("Should create a RAG layout with valid code", () => {
             const refField = LayoutBuilder.create("TEXT");
             refField.withName("refField").withShortName(refShortName);
-            const builder = LayoutBuilder.create("RAG");
+            const builder:ICodeLayoutBuilder = LayoutBuilder.create("RAG") as ICodeLayoutBuilder;
             builder
                 .withName("test1")
                 .withShortName("t1")
@@ -80,7 +81,7 @@ end
         it("Should error on creating a CALC layout with invalid code", () => {
             const refField = LayoutBuilder.create("TEXT");
             refField.withName("refField").withShortName(refShortName);
-            const builder = LayoutBuilder.create("CALC");
+            const builder:ICodeLayoutBuilder = LayoutBuilder.create("CALC") as ICodeLayoutBuilder;
             builder
                 .withName("test1")
                 .withShortName("t1")
@@ -101,7 +102,7 @@ end
         it("Should create a CALC layout with valid code", () => {
             const refField = LayoutBuilder.create("TEXT");
             refField.withName("refField").withShortName(refShortName);
-            const builder = LayoutBuilder.create("CALC");
+            const builder:ICodeLayoutBuilder = LayoutBuilder.create("CALC") as ICodeLayoutBuilder;
             builder
                 .withName("test1")
                 .withShortName("t1")
@@ -123,7 +124,7 @@ end
 
     context("Layout creation with dropdown", () => {
         it("Creates a dropdown", () => {
-            const builder = <IDropdownLayoutBuilder>LayoutBuilder.create("DROPDOWN");
+            const builder:IDropdownLayoutBuilder = LayoutBuilder.create("DROPDOWN") as IDropdownLayoutBuilder;
             builder
                 .withName("test1")
                 .withShortName("t1")
@@ -154,7 +155,7 @@ end
             // Go to table1
             cy.gotoInstanceByShortName("table1", "layout")
             // Create curval field with reference to reference field
-            const builder = <ICurvalLayoutBuilder>LayoutBuilder.create("CURVAL");
+            const builder:ICurvalLayoutBuilder = LayoutBuilder.create("CURVAL") as ICurvalLayoutBuilder;
             builder
                 .withName("test1")
                 .withShortName("t1")
