+++ cypress/support/builders/layout/types.ts
@@ -1,0 +1,1 @@
+export type LayoutType = "TEXT" | "INTEGER" | "DATE" | "DATE-RANGE" | "DROPDOWN" | "TREE" | "DOCUMENT" | "PERSON" | "RAG" | "CALC" | "CURVAL" | "AUTOCUR";
