+++ cypress/support/constants.ts
@@ -1,0 +1,4 @@
+export const goodUser = "test@example.com";
+export const goodPassword = "xyz123";
+
+export type instanceMode = "layout" | "permissions" | "topics" | "edit" | "data" | "report" | undefined;
